refactor: remove GSTR-1 and rename GSTR-1 Beta to GSTR-1
