+++ india_compliance/gst_india/doctype/gst_return_log/gst_return_log.py
@@ -306,8 +306,8 @@ def update_is_not_latest_gstr1_data(posting_date, company_gstin):
     frappe.publish_realtime(
         "is_not_latest_gstr1_data",
         message={"filters": {"company_gstin": company_gstin, "period": period}},
-        doctype="GSTR-1 Beta",
-        docname="GSTR-1 Beta",
+        doctype="GSTR-1",
+        docname="GSTR-1",
     )
