+++ india_compliance/gst_india/doctype/gstr_1/gstr_1.json
@@ -110,7 +110,7 @@
  "modified": "2025-02-06 12:26:08.792234",
  "modified_by": "Administrator",
  "module": "GST India",
- "name": "GSTR-1 Beta",
+ "name": "GSTR-1",
  "owner": "Administrator",
  "permissions": [
   {
