+++ india_compliance/gst_india/doctype/gstr_1/test_gstr_1.py
@@ -5,5 +5,5 @@
 from frappe.tests import IntegrationTestCase
 
 
-class TestGSTR1Beta(IntegrationTestCase):
+class TestGSTR1(IntegrationTestCase):
     pass
