--- india_compliance/gst_india/report/gstr_1/gstr_1.json
@@ -1,32 +1,0 @@
-{
- "add_total_row": 1,
- "columns": [],
- "creation": "2018-01-02 15:54:41.424225",
- "disable_prepared_report": 0,
- "disabled": 0,
- "docstatus": 0,
- "doctype": "Report",
- "filters": [],
- "idx": 0,
- "is_standard": "Yes",
- "modified": "2022-03-01 15:46:30.974086",
- "modified_by": "Administrator",
- "module": "GST India",
- "name": "GSTR-1",
- "owner": "Administrator",
- "prepared_report": 0,
- "ref_doctype": "GL Entry",
- "report_name": "GSTR-1",
- "report_type": "Script Report",
- "roles": [
-  {
-   "role": "Accounts Manager"
-  },
-  {
-   "role": "Accounts User"
-  },
-  {
-   "role": "Auditor"
-  }
- ]
-}
