+++ india_compliance/gst_india/utils/__init__.py
@@ -1089,3 +1089,24 @@ def enable_autocommit(fn):
             db.auto_commit_on_many_writes = autocommit
 
     return wrapper
+
+
+def get_company_gstin_number(company, address=None, all_gstins=False):
+    gstin = ""
+    if address:
+        gstin = frappe.db.get_value("Address", address, "gstin")
+
+    if not gstin:
+        gstin = get_gstin_list(company)
+        if gstin and not all_gstins:
+            gstin = gstin[0]
+
+    if not gstin:
+        address = frappe.bold(address) if address else ""
+        frappe.throw(
+            _("Please set valid GSTIN No. in Company Address {} for company {}").format(
+                address, frappe.bold(company)
+            )
+        )
+
+    return gstin
