+++ india_compliance/patches/v15/migrate_gstr1_log_to_returns_log.py
@@ -17,5 +17,5 @@ def execute():
     # Drop the old table
     frappe.db.delete("GSTR-1 Log")
 
-    # Clear all fields saved in GSTR-1 Beta
-    frappe.db.delete("Singles", {"doctype": "GSTR-1 Beta"})
+    # Clear all fields saved in GSTR-1
+    frappe.db.delete("Singles", {"doctype": "GSTR-1"})
