+++ india_compliance/public/js/help_links.js
@@ -15,11 +15,11 @@ frappe.help.help_links["india-compliance-account"] = [
 frappe.help.help_links["Form/GST Settings"] = [
     {
         label: "Setting Up GST accounts",
-        url: docsUrl + "configuration/gst_setup#gst-accounts"
+        url: docsUrl + "configuration/gst_setup#gst-accounts",
     },
     {
         label: "Setting Up API",
-        url: docsUrl + "ewaybill-and-einvoice/gst_settings"
+        url: docsUrl + "ewaybill-and-einvoice/gst_settings",
     },
 ];
 
@@ -28,9 +28,8 @@ frappe.help.help_links["Form/Company"] = [
     {
         label: "Print Settings",
         url: docsUrl + "configuration/gst_setup#print-format",
-    }
+    },
 ];
-
 
 //Doctypes
 //Sales Invoice
@@ -46,14 +45,14 @@ frappe.help.help_links["Form/Sales Invoice"].push(
     {
         label: "e-Invoice",
         url: docsUrl + "ewaybill-and-einvoice/generating_e_invoice",
-    },
+    }
 );
 
 //Stock Entry
 frappe.help.help_links["Form/Stock Entry"].push({
     label: "Subcontracting Workflow",
     url: blogUrl + "posts/post5",
-})
+});
 
 //Subcontracting Receipt
 frappe.help.help_links["Form/Subcontracting Receipt"] = [
@@ -63,22 +62,24 @@ frappe.help.help_links["Form/Subcontracting Receipt"] = [
     },
     {
         label: "GST Job Work Stock Movement report",
-        url: docsUrl + "gst-reports/miscellaneous_reports#gst-job-work-stock-movement-report",
+        url:
+            docsUrl +
+            "gst-reports/miscellaneous_reports#gst-job-work-stock-movement-report",
     },
-]
+];
 
 //Journal Entry
 frappe.help.help_links["Form/Journal Entry"] = [
     {
         label: "Reversal of Input Tax Credit",
         url: docsUrl + "configuration/other_transaction#reversal-of-input-tax-credit",
-    }
-]
+    },
+];
 
 // GST Reports
-frappe.help.help_links["Form/GSTR-1 Beta"] = [
+frappe.help.help_links["Form/GSTR-1"] = [
     {
-        label: "GSTR-1 Beta",
+        label: "GSTR-1",
         url: docsUrl + "gst-reports/gstr1",
     },
 ];
@@ -97,12 +98,13 @@ frappe.help.help_links["List/GSTR 3B Report"] = [
     },
 ];
 
-
 //Query Reports
 frappe.help.help_links["query-report/GST Job Work Stock Movement"] = [
     {
         label: "GST Job Work Stock Movement",
-        url: docsUrl + "gst-reports/miscellaneous_reports#gst-job-work-stock-movement-report",
+        url:
+            docsUrl +
+            "gst-reports/miscellaneous_reports#gst-job-work-stock-movement-report",
     },
 ];
 
@@ -116,14 +118,18 @@ frappe.help.help_links["query-report/GST Balance"] = [
 frappe.help.help_links["query-report/GST Sales Register Beta"] = [
     {
         label: "GST Sales Register Beta",
-        url: docsUrl + "gst-reports/miscellaneous_reports#gst-sales-register-beta-report",
+        url:
+            docsUrl +
+            "gst-reports/miscellaneous_reports#gst-sales-register-beta-report",
     },
 ];
 
 frappe.help.help_links["query-report/GST Purchase Register"] = [
     {
         label: "GST Purchase Register",
-        url: docsUrl + "gst-reports/miscellaneous_reports#gst-purchase-register-beta-report",
+        url:
+            docsUrl +
+            "gst-reports/miscellaneous_reports#gst-purchase-register-beta-report",
     },
 ];
